cleanup and new recievers.py for the Reciever class
